Debug1,2,3,6, and an additional issue
Debugging roller coaster simulator:

Log by Zhilin:

Issue 6: There are times when the output values do not fit into the boxes, like for the value for acceleration here.
--Change the position of buttons “reset” and “Delete Track” to the right.
--Make the panels “Inputs”, “Forces”, and “Speed” larger in width.


Issue 3: Roller coaster car is too large relative to the track. Want the car size changes with the mass more obviously.
--Make the car change more obviously

Issue 1:  the car and top of the initial drop are not visible on the screen (you can see it if the screen size is minimized, but seems problematic.)
Issue 2:  Relatedly, at one point, I clicked on “delete track” I was able to delete several parts of the track. But as you can see, I was not able to delete the initial drop, which is almost completely off the screen (see in screenshot on right).
--Track.js: added two methods: getLowestX() and bumpAsideLeftWindow(), which ensures the track not going out of the left limit of window
--SimulationModel.js: call newTrack.bumpAsideLeftWindow() twice
--ControlPointNode.js: call track.bumpAsideLeftWindow()
--TrackDragHandler.js: call track.bumpAsideLeftWindow()

New Issue: What if the track goes out of the right limit?
--Track.js: added two methods: getHighestX() and bumpAsideRightWindow(), which ensures the track not going out of the right limit of window
--SimulationModel.js: call newTrack.bumpAsideRightWindow() twice
--ControlPointNode.js: call track.bumpAsideRightWindow()
--TrackDragHandler.js: call track.bumpAsideRightWindow()